requirements.txt
# Streamlit app + CPU TensorFlow
streamlit==1.51.0

# Use CPU-only TensorFlow compatible with many minimal containers
tensorflow-cpu==2.12.0

# Match scikit-learn version used when models were likely saved
scikit-learn==1.6.1

# Pin protobuf to avoid TF/protobuf runtime issues
protobuf==3.20.*

# common packages used by the app
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.7
altair==5.5.0
joblib==1.5.2
h5py==3.8.0
pyarrow==21.0.0

# extras you might need
jinja2==3.1.6
requests==2.32.5
tqdm==4.66.1
python-dateutil==2.9.0.post0
